# 02_Server/swagger.json
{
  "swagger": "2.0",
  "info": {
  "version": "1.0.0",
    "title": "Sample spec",
    "description": "Sample spec for Swagger",
    "termsOfService": "http://swagger.io/terms/"
},
  "basePath": "/",
  "tags" : [
  {"name": "Tag1",
    "description": "API for something"
  }
],
  "schemes": [
  "https"
],
  "consumes": [
  "application/json"
],
  "produces": [
  "application/json"
],
  "paths": {
  "/path/to/method": {
    "post": {
      "tags": ["Tag1"],
        "description": "what this API does",
        "operationId": "for the URL",
        "produces": [
        "application/json"
      ],
        "parameters": [
        {
          "name": "Job Parameters",
          "in": "body",
          "description": "Job to add to the database",
          "required": true,
          "schema": {
            "$ref": "#/definitions/create"
          }
        }
      ],
        "responses": {
        "200": {
          "description": "create response",
            "schema": {
            "$ref": "#/definitions/response"
          }
        }
      }
    }
  },
  "/path/to/method/{someVariable}": {
    "delete":{
      "tags":["Tag1"],
        "description":"Deletes a something",
        "operationId":"delete",
        "produces": [
        "application/json"
      ],
        "parameters":[
        {
          "name": "someVariable",
          "in": "path",
          "description": "ID to delete",
          "required": true,
          "type": "string"
        }
      ],
        "responses": {
        "200": {
          "description": "delete response",
            "schema":{
            "$ref":"#/definitions/deleteRes"
          }
        }
      }
    },
    "get":{
      "tags": [ "Tag1"],
        "description": "Returns the status of something",
        "operationId": "findById",
        "produces": [
        "application/json"
      ],
        "parameters": [
        {
          "name": "ID",
          "in": "path",
          "description": "ID to fetch",
          "required": true,
          "type": "string"
        }
      ],
        "responses": {
        "200": {
          "description": "response",
            "schema": {
            "$ref": "#/definitions/response1"
          }
        }
      }
    }
  }
},
  "definitions": {
  "create": {
    "properties": {
      "eventId": {
        "required":true,
          "type": "string",
          "example": "mathmaticial"
      },
      "numberRequired": {
        "required":true,
          "type":"integer",
          "format": "int64",
          "example": 500
      }
    }
  },
  "response1": {
    "properties": {
      "jobID": {
        "required":true,
          "type": "string",
          "example": "23423456543"
      },
      "request": {
        "required":true,
          "type": "schema",
          "example": {
          "eventId":"mathmaticial",
            "numberRequired": 500
        }
      }
    }
  },
  "response": {
    "properties": {
      "jobID": {
        "required": true,
          "type" : "string",
          "example" : "12321432423"
      }
    }
  },
  "delete": {
    "properties": {
      "jobID": {
        "required": true,
          "type" : "string",
          "example" : "234344234234"
      },
      "status" : {
        "required": true,
          "type": "string",
          "example": "deleted"
      }
    }
  },
  "400error" : {
    "properties": {
      "message": {
        "type": "string",
          "example" : "400 Bad Request"
      }
    }
  }
}
}